rename airview-store directory
