+++ packages/airview-cms/src/features/index.js
@@ -1,4 +1,4 @@
-export * from "./airview-store";
+export * from "./store";
 export * from "./cms";
 export * from "./use-get-all-entries-meta";
 export * from "./use-get-entry";
