+++ packages/airview-cms/src/features/store/airview-provider.js
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { Provider } from "react-redux";
-import { airviewStore } from "./airview-store";
+import { store } from "./store";
 
 export function AirviewProvider({ children }) {
-  return <Provider store={airviewStore}>{children}</Provider>;
+  return <Provider store={store}>{children}</Provider>;
 }
 
 AirviewProvider.propTypes = {
