+++ packages/airview-cms/src/views/main-view/main-view.js
@@ -21,17 +21,3 @@ export function MainView({ children }) {
 MainView.propTypes = {
   children: PropTypes.node,
 };
-
-/*
-- cms
-- - index.js
-- - cms.js
-- - cms.slice.js
-- - main-content
-- - enable-cms-button
-- - cms-loading-indicator
-- - branch-creator
-- - meta-editor
-- - toolbar
-- - use-set-cms-context
-*/
